правки с прошлого задания
