+++ js/main.js
@@ -3,8 +3,9 @@ import './map.js';
 import './user-form.js';
 import { resetToDeafaultState } from './page-state.js'
 import { setUserFormSubmit } from './user-form.js';
+import { initMap } from './map.js';
 
-
+initMap()
 setUserFormSubmit(resetToDeafaultState);
