+++ js/map-filter.js
@@ -1,0 +1,9 @@
+const getFilteredAds = (ads, typeHouses) => {
+  let filteredAds = ads;
+  if (typeHouses) {
+    filteredAds = ads.filter(ad => ad.offer.type.includes(typeHouses));
+  }
+  return filteredAds;
+}
+
+export { getFilteredAds }
